Moved RPC objects out of RPCClient file.
