+++ src/Stratis.Bitcoin.Features.RPC/ChangeAddress.cs
@@ -1,0 +1,10 @@
+﻿using NBitcoin;
+
+namespace Stratis.Bitcoin.Features.RPC
+{
+    public class ChangeAddress
+    {
+        public Money Amount { get; set; }
+        public BitcoinAddress Address { get; set; }
+    }
+}
