+++ src/Stratis.Bitcoin.Features.RPC/Exceptions/RPCServerException.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using NBitcoin.RPC;
 
-namespace Stratis.Bitcoin.Features.RPC
+namespace Stratis.Bitcoin.Features.RPC.Exceptions
 {
     public class RPCServerException : Exception
     {
