+++ src/Stratis.Bitcoin.Features.RPC/NodeAddressInfo.cs
@@ -1,0 +1,10 @@
+﻿using System.Net;
+
+namespace Stratis.Bitcoin.Features.RPC
+{
+    public class NodeAddressInfo
+    {
+        public IPEndPoint Address { get; internal set; }
+        public bool Connected { get; internal set; }
+    }
+}
