+++ src/Stratis.Bitcoin.Features.RPC/RPCAccount.cs
@@ -1,0 +1,10 @@
+﻿using NBitcoin;
+
+namespace Stratis.Bitcoin.Features.RPC
+{
+    public class RPCAccount
+    {
+        public Money Amount { get; set; }
+        public string AccountName { get; set; }
+    }
+}
