+++ src/Stratis.Bitcoin.Features.RPC/RPCClient.Wallet.cs
@@ -9,31 +9,6 @@ using Newtonsoft.Json.Linq;
 
 namespace Stratis.Bitcoin.Features.RPC
 {
-    public class RPCAccount
-    {
-        public Money Amount { get; set; }
-        public string AccountName { get; set; }
-    }
-
-    public class ChangeAddress
-    {
-        public Money Amount { get; set; }
-        public BitcoinAddress Address { get; set; }
-    }
-
-    public class AddressGrouping
-    {
-        public AddressGrouping()
-        {
-            this.ChangeAddresses = new List<ChangeAddress>();
-        }
-
-        public BitcoinAddress PublicAddress { get; set; }
-        public Money Amount { get; set; }
-        public string Account { get; set; }
-        public List<ChangeAddress> ChangeAddresses { get; set; }
-    }
-
     /*
         Category            Name                        Implemented 
         ------------------ --------------------------- -----------------------
