+++ src/Stratis.Bitcoin.Features.RPC/RPCClient.cs
@@ -12,6 +12,7 @@ using NBitcoin.DataEncoders;
 using NBitcoin.RPC;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using Stratis.Bitcoin.Features.RPC.Exceptions;
 
 namespace Stratis.Bitcoin.Features.RPC
 {
@@ -1346,53 +1347,6 @@ namespace Stratis.Bitcoin.Features.RPC
         public uint256[] Generate(int nBlocks)
         {
             return GenerateAsync(nBlocks).GetAwaiter().GetResult();
-        }
-    }
-
-    public class PeerInfo
-    {
-        public int Id { get; internal set; }
-        public IPEndPoint Address { get; internal set; }
-        public IPEndPoint LocalAddress { get; internal set; }
-        public ulong Services { get; internal set; }
-        public DateTimeOffset LastSend { get; internal set; }
-        public DateTimeOffset LastReceive { get; internal set; }
-        public long BytesSent { get; internal set; }
-        public long BytesReceived { get; internal set; }
-        public DateTimeOffset ConnectionTime { get; internal set; }
-        public TimeSpan? PingTime { get; internal set; }
-        public int Version { get; internal set; }
-        public string SubVersion { get; internal set; }
-        public bool Inbound { get; internal set; }
-        public int StartingHeight { get; internal set; }
-        public int BanScore { get; internal set; }
-        public int SynchronizedHeaders { get; internal set; }
-        public int SynchronizedBlocks { get; internal set; }
-        public uint[] Inflight { get; internal set; }
-        public bool IsWhiteListed { get; internal set; }
-        public TimeSpan PingWait { get; internal set; }
-        public int Blocks { get; internal set; }
-        public TimeSpan TimeOffset { get; internal set; }
-    }
-
-    public class AddedNodeInfo
-    {
-        public EndPoint AddedNode { get; internal set; }
-        public bool Connected { get; internal set; }
-        public IEnumerable<NodeAddressInfo> Addresses { get; internal set; }
-    }
-
-    public class NodeAddressInfo
-    {
-        public IPEndPoint Address { get; internal set; }
-        public bool Connected { get; internal set; }
-    }
-
-    public class NoEstimationException : Exception
-    {
-        public NoEstimationException(int nblock)
-            : base("The FeeRate couldn't be estimated because of insufficient data from Bitcoin Core. Try to use smaller nBlock, or wait Bitcoin Core to gather more data.")
-        {
         }
     }
 }
